Delete button manage + remove bli giver button from bedrift page (#820)
* close button municipality editor

* close button municipality editor

* Navbar landing page (#781)

* navbar start

* start

* moved navbar and button

* giver button top right

* navbar position and color

* added logo to navbar, dynamic positioning

* added more padding

* drawer for mobile view

* Managment Api created, and route for user creating added. (#813)

* okey

* Update GiEnJul/Dtos/CreateUserDto.cs



* Manage municipalities (#807)

* all backend func for municipality table

* start component municipality

* edit

* added methods backend and edit component

* start component municipality

* working but ugly

* initstate and remove list

* fix

* removed rowKey from controller and rebase fix

* added email to municipality table

* confirmation frontend and edit method backend

* all backend func for municipality table

* start component municipality

* edit

* added methods backend and edit component

* start component municipality

* working but ugly

* initstate and remove list

* fix

* removed rowKey from controller and rebase fix

* added email to municipality table

* confirmation frontend and edit method backend

* added mapping for Email in table

* all backend func for municipality table

* added methods backend and edit component

* start component municipality

* working but ugly

* initstate and remove list

* fix

* removed rowKey from controller and rebase fix

* added email to municipality table

* confirmation frontend and edit method backend

* added mapping for Email in table

* Update GiEnJul/Controllers/MunicipalityController.cs


* Update GiEnJul/Controllers/MunicipalityController.cs


* Update GiEnJul/Controllers/MunicipalityController.cs


* test

* post municipality method fix

* delete unused method


* close button municipality editor

* del button, add link, clean up navbar

* remove const tab
